src/main/webapp/app/entities/fornitore/route/fornitore-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IFornitore, Fornitore } from '../fornitore.model';
import { FornitoreService } from '../service/fornitore.service';

import { FornitoreRoutingResolveService } from './fornitore-routing-resolve.service';

describe('Fornitore routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: FornitoreRoutingResolveService;
  let service: FornitoreService;
  let resultFornitore: IFornitore | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(FornitoreRoutingResolveService);
    service = TestBed.inject(FornitoreService);
    resultFornitore = undefined;
  });

  describe('resolve', () => {
    it('should return IFornitore returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultFornitore = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultFornitore).toEqual({ id: 123 });
    });

    it('should return new IFornitore if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultFornitore = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultFornitore).toEqual(new Fornitore());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Fornitore })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultFornitore = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultFornitore).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
